wip: attestation service and CPI for schema registry
