+++ contracts/solana/programs/attestation/src/sdk.rs
@@ -1,0 +1,50 @@
+// Define the external program's interface
+pub mod external_resolver {
+    use anchor_lang::prelude::*;
+    use super::*;
+    
+    #[derive(Accounts)]
+    pub struct ExternalInstruction<'info> {
+        // Define necessary accounts
+        pub account1: AccountInfo<'info>,
+        pub account2: AccountInfo<'info>,
+        // ... other accounts
+    }
+    
+    pub fn external_instruction(ctx: Context<ExternalInstruction>, arg1: u64, arg2: String) -> Result<()> {
+        // Implementation not needed; this is just to define the interface
+        Ok(())
+    }
+}
+
+pub fn call_external_program(
+    ctx: Context<CallExternalProgram>,
+    arg1: u64,
+    arg2: String,
+) -> Result<()> {
+    // Serialize the instruction data
+    let instruction_data = external_program::instruction::ExternalInstruction { arg1, arg2 };
+    
+    // Define the external program's ID
+    let external_program_id = ctx.accounts.external_program.key();
+    
+    // Define the accounts required by the external program
+    let account1 = ctx.accounts.account1.to_account_info();
+    let account2 = ctx.accounts.account2.to_account_info();
+    // ... other accounts
+    
+    // Create the CPI context
+    let cpi_ctx = CpiContext::new(
+        ctx.accounts.external_program.to_account_info(),
+        external_program::instruction::ExternalInstruction {
+            account1: account1.clone(),
+            account2: account2.clone(),
+            // ... other accounts
+        },
+    );
+    
+    // Invoke the external program's instruction
+    external_program::external_instruction(cpi_ctx, arg1, arg2)?;
+    
+    Ok(())
+}
