+++ contracts/solana/programs/authority/Cargo.toml
@@ -17,5 +17,5 @@ no-log-ix-name = []
 idl-build = ["anchor-lang/idl-build"]
 
 [dependencies]
-anchor-lang = { version = "0.30.1", features = ["init-if-needed"] }
+anchor-lang = { version = "0.30.1" }
 solana-security-txt = "1.1.1"
