+++ contracts/solana/programs/registry/Cargo.toml
@@ -17,4 +17,5 @@ no-log-ix-name = []
 idl-build = ["anchor-lang/idl-build"]
 
 [dependencies]
-anchor-lang = "0.30.1"
+anchor-lang = { version = "0.30.1", features = ["init-if-needed"] }
+solana-security-txt = "1.1.1"
